requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pymongo
python-dotenv
jinja2
python-multipart
markdown
python-jose
auth0-python